THIS IS NOT THE FINAL VERSION
Created leaflet -> I've just finished the first version of it.it might do the jobs but some points are still unclear as an example what is a leaflet.
I've searched on the Internet and found the leaflet is an image or a text so I've implemented both on the page.
There's also security issues like the picture size or if it's already in the database.
By the way I've also added some to do things on the list.
But I will explain it tomorrow during the class